MR : Fix: Mod VoteRepository(countByQuestionGroupAndQuestionType,... | controller => deploy-test

See merge request s09-webmobile3-sub2/S09P12A304!140

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/response/AllAnswerResponse.java
@@ -22,8 +22,8 @@ public class AllAnswerResponse {
 	private Long answerA;
 	private Long answerB;
 	// 비율
-	private double rateA;
-	private double rateB;
+	private Long rateA;
+	private Long rateB;
 
 	public AllAnswerResponse(Question entity) {
 		this.questionID = entity.getQuestionID();
